+++ jhm/config.h
@@ -122,7 +122,7 @@ class TConfig final {
   void AddFile(const std::string &filename);
 
   Util::TOptTimestamp Timestamp;
-  Base::TOpt<uint64_t> ComputedStart;
+  Base::TOpt<ssize_t> ComputedStart;
   mutable bool ComputedLocked =
       false;  // This is purely an internal safety check / correctness check
   std::deque<Base::TJson> Config;
